Make InMemoryTermIdsAcquirer implement TermCleaner
The TermIdsAcquirer and the TermCleaner clearly have to manipulate the
same state somehow. For the MediaWiki implementation, this happens via
the database; for the in-memory implementation, I’m going the easy route
and making one class implement both interfaces (rather than, for
example, having a third class with the actual state – the $terms array –
and two classes accessing that and implementing the acquirer and cleaner
parts). Some of the tests could be reused for other cleaners as well in
the future, if we wanted.