Improve testCleanTerms_completelyCleansArray
The two calls to acquireTermIds seemed to be accidental
complexity so I changed to only one call.

I also removed access to the private field as this is surprising
and breaks automated refactoring. New approach has class define
explicitly what can be accessed.

... Just realized:

Since the tested class is an in memory implementation meant
for tests and implements a package private interface, it is
probably better placed in tests/TestDoubles or similar rather
than src/. Lets clean that up in some follow up.