Add DatabaseTermIdsResolver
This is the database implementation of the TermIdsResolver interface,
the inverse of DatabaseTermIdsAcquirer. For now, the implementation is
fairly simple and uses a plain JOIN for most fields (though types
already go through a TypeIdResolver); we’ll probably want to introduce
caching (on several layers) in the future.

The test inserts are based on MediaWikiNormalizedTermCleanerTest.