Inject ILoadBalancer instead of IDatabase instances
Injecting IDatabase instances means that a database connection must
already be open just to instantiate the object (unless DBConnRef is
used), which forces users to lazily instantiate the object unless they
want to incur the cost of a database connection on every request. It’s
better to inject an ILoadBalancer instead and defer acquiring the
connection until we really need it; in fact, in some cases we may not
need a DB_MASTER connection at all, and might instead be able to make do
with just a DB_REPLICA connection.

A test helper class is included that can be used instead of a real load
balancer to return one or two pre-configured connection objects.

(Also adds the missing declaration of DatabaseTermIdsAcquirer’s
$typeIdsAqcuirer, while we’re at it.)