Rename in-memory implementations from Acquirer to Store
Both of them now implement more interfaces than just an Acquirer, so
let’s make the name a bit more generic to reflect that.

In InMemoryTermIdsStoreTest, variables for an instance of this class are
also renamed, but only where they actually use more than one interface.
The tests for the Acquirer part don’t rely on the test of the store and
could in future be refactored to test other Acquirer implementations as
well, so let’s leave those Acquirer-specific for now.